Renamed PropertyCard to PropertyCardAdmin and created PropertyCard component to be only rendered in propertylistingpage
