+++ src/components/PropertyCardAdmin.jsx
@@ -1,10 +1,10 @@
-import React, { useContext } from 'react'
-import { Link } from 'react-router-dom'
-import { BsArrowsFullscreen, BsPeople } from 'react-icons/bs'
-import { UserContext } from '../context/userContext' // Assuming you have a UserContext to manage user state
+import React, { useContext } from 'react';
+import { Link } from 'react-router-dom';
+import { BsArrowsFullscreen, BsPeople } from 'react-icons/bs';
+import { UserContext } from '../context/userContext'; // Assuming you have a UserContext to manage user state
 
-const PropertyCard = ({ property, onDelete, onEdit }) => {
-  const { user } = useContext(UserContext) // Get user context
+const PropertyCardAdmin = ({ property, onDelete, onEdit }) => {
+  const { user } = useContext(UserContext); // Get user context to check if the user is an admin
 
   return (
     <div className='bg-white shadow-2xl min-h-[500px] group'>
@@ -19,7 +19,7 @@ const PropertyCard = ({ property, onDelete, onEdit }) => {
                 alt={property.name}
               />
             ) : (
-              <div className='w-full h-48 bg-gray-200 flex justify-center items-center'>
+              <div className="w-full h-48 bg-gray-200 flex justify-center items-center">
                 <span>No Image Available</span>
               </div>
             )}
@@ -58,20 +58,37 @@ const PropertyCard = ({ property, onDelete, onEdit }) => {
               {property.description.slice(0, 56)}...
             </p>
           </div>
-          
-          {/* Booking Link for all users */}
-          <Link
-            to={`/booking/${property._id}`}
-            className='btn btn-secondary btn-sm max-w-[240px] mx-auto'
-          >
-            Book now from ${property.price}
-          </Link>
+
+          {/* Conditional Rendering for Admins and Users */}
+          {user && user.isAdmin ? (
+            <div className="flex flex-col gap-2 p-4">
+              <button
+                onClick={() => onEdit(property)}
+                className="btn btn-secondary btn-sm max-w-[240px] mx-auto"
+              >
+                Edit Property
+              </button>
+              <button
+                onClick={() => onDelete(property._id)}
+                className="btn btn-secondary btn-sm max-w-[240px] mx-auto bg-red-500 hover:bg-red-600"
+              >
+                Delete Property
+              </button>
+            </div>
+          ) : (
+            <Link
+              to={`/booking/${property._id}`}
+              className='btn btn-secondary btn-sm max-w-[240px] mx-auto'
+            >
+              Book now from ${property.price}
+            </Link>
+          )}
         </>
       ) : (
         <p>Loading property details...</p>
       )}
     </div>
-  )
-}
+  );
+};
 
-export default PropertyCard
+export default PropertyCardAdmin;
